# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Empresas",
        "description": "",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Entorno de Producción (Consume Créditos)"
        },
        {
            "url": "/api/sandbox/v1",
            "description": "Entorno Sandbox de Pruebas (Gratis. Usa CIF: A15075062)"
        }
    ],
    "paths": {
...
        "/api/v1/companies/network": {
            "get": {
                "tags": [
                    "2. Plan Pro"
                ],
                "summary": "Grafos de Poder Societario",
                "description": "Obtiene la red de vinculación entre empresas a través de sus administradores. **Coste:** 1 llamada de tu cuota mensual (plan suscripción) o 3 créditos del monedero (bono prepago). Las respuestas con error no consumen cuota ni créditos.",
                "operationId": "b5b7a1477c590925db6c3751ca3660ec",
                "parameters": [
                    {
                        "name": "cif",
                        "in": "query",
                        "description": "CIF de la empresa base",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grafo de nodos y conexiones (empresas y administradores)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/professional/search": {
            "get": {
                "tags": [
                    "2. Plan Pro"
                ],
                "summary": "Búsqueda Autocompletado Profesional",
                "description": "Devuelve coincidencias ligeras para autocompletado en búsquedas.",
                "operationId": "a2461cbe092e786068a34d6fcc0d8eae",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados del autocompletado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/professional/details": {
            "get": {
                "tags": [
                    "2. Plan Pro"
                ],
                "summary": "Detalles Profesionales de Empresa",
                "description": "Devuelve los datos profesionales de una empresa basados en su CIF.",
                "operationId": "508a95024c235c9cc833d09e22138480",
                "parameters": [
                    {
                        "name": "cif",
                        "in": "query",
                        "description": "El CIF de la empresa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles completos de la empresa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Introduce tu API Key privada aquí para autenticarte y poder probar los endpoints en tiempo real.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
...
}